Rename BoxControl's local input component to avoid InputControl confusion.
